Cleanup and restructure Pydantic models for Schema
