+++ infrahub_sdk/schema/__init__.py
@@ -4,35 +4,62 @@ import asyncio
 from collections import defaultdict
 from collections.abc import MutableMapping
 from enum import Enum
-from pathlib import Path
 from time import sleep
-from typing import TYPE_CHECKING, Any, Optional, TypedDict, TypeVar, Union
+from typing import TYPE_CHECKING, Any, Optional, TypedDict, Union
 from urllib.parse import urlencode
 
 import httpx
-from pydantic import BaseModel, ConfigDict, Field, field_validator
+from pydantic import BaseModel, Field
 from typing_extensions import TypeAlias
 
-from ._importer import import_module
-from .checks import InfrahubCheck
-from .exceptions import (
+from ..exceptions import (
     InvalidResponseError,
-    ModuleImportError,
-    ResourceNotDefinedError,
     SchemaNotFoundError,
     ValidationError,
 )
-from .generator import InfrahubGenerator
-from .graphql import Mutation
-from .queries import SCHEMA_HASH_SYNC_STATUS
-from .transforms import InfrahubTransform
-from .utils import duplicates
+from ..graphql import Mutation
+from ..queries import SCHEMA_HASH_SYNC_STATUS
+from .main import (
+    AttributeSchema,
+    AttributeSchemaAPI,
+    GenericSchema,
+    GenericSchemaAPI,
+    NodeSchema,
+    NodeSchemaAPI,
+    ProfileSchemaAPI,
+    RelationshipCardinality,
+    RelationshipKind,
+    RelationshipSchema,
+    RelationshipSchemaAPI,
+    SchemaRoot,
+    SchemaRootAPI,
+)
+from .repository import InfrahubRepositoryConfig
 
 if TYPE_CHECKING:
-    from .client import InfrahubClient, InfrahubClientSync, SchemaType, SchemaTypeSync
-    from .node import InfrahubNode, InfrahubNodeSync
+    from ..client import InfrahubClient, InfrahubClientSync, SchemaType, SchemaTypeSync
+    from ..node import InfrahubNode, InfrahubNodeSync
 
     InfrahubNodeTypes = Union[InfrahubNode, InfrahubNodeSync]
+
+
+__all__ = [
+    "AttributeSchema",
+    "AttributeSchemaAPI",
+    "GenericSchema",
+    "GenericSchemaAPI",
+    "InfrahubRepositoryConfig",
+    "NodeSchema",
+    "NodeSchemaAPI",
+    "ProfileSchemaAPI",
+    "RelationshipCardinality",
+    "RelationshipKind",
+    "RelationshipSchema",
+    "RelationshipSchemaAPI",
+    "SchemaRoot",
+    "SchemaRootAPI",
+]
+
 
 # pylint: disable=redefined-builtin
 
@@ -41,267 +68,6 @@ class DropdownMutationOptionalArgs(TypedDict):
     color: Optional[str]
     description: Optional[str]
     label: Optional[str]
-
-
-ResourceClass = TypeVar("ResourceClass")
-
-# ---------------------------------------------------------------------------------
-# Repository Configuration file
-# ---------------------------------------------------------------------------------
-
-
-class InfrahubRepositoryConfigElement(BaseModel):
-    """Class to regroup all elements of the infrahub configuration for a repository for typing purpose."""
-
-
-class InfrahubRepositoryArtifactDefinitionConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the artifact definition")
-    artifact_name: Optional[str] = Field(default=None, description="Name of the artifact created from this definition")
-    parameters: dict[str, Any] = Field(..., description="The input parameters required to render this artifact")
-    content_type: str = Field(..., description="The content type of the rendered artifact")
-    targets: str = Field(..., description="The group to target when creating artifacts")
-    transformation: str = Field(..., description="The transformation to use.")
-
-
-class InfrahubJinja2TransformConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the transform")
-    query: str = Field(..., description="The name of the GraphQL Query")
-    template_path: Path = Field(..., description="The path within the repository of the template file")
-    description: Optional[str] = Field(default=None, description="Description for this transform")
-
-    @property
-    def template_path_value(self) -> str:
-        return str(self.template_path)
-
-    @property
-    def payload(self) -> dict[str, str]:
-        data = self.model_dump(exclude_none=True)
-        data["template_path"] = self.template_path_value
-        return data
-
-
-class InfrahubCheckDefinitionConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the Check Definition")
-    file_path: Path = Field(..., description="The file within the repository with the check code.")
-    parameters: dict[str, Any] = Field(
-        default_factory=dict, description="The input parameters required to run this check"
-    )
-    targets: Optional[str] = Field(
-        default=None, description="The group to target when running this check, leave blank for global checks"
-    )
-    class_name: str = Field(default="Check", description="The name of the check class to run.")
-
-    def load_class(self, import_root: Optional[str] = None, relative_path: Optional[str] = None) -> type[InfrahubCheck]:
-        module = import_module(module_path=self.file_path, import_root=import_root, relative_path=relative_path)
-
-        if self.class_name not in dir(module):
-            raise ModuleImportError(message=f"The specified class {self.class_name} was not found within the module")
-
-        check_class = getattr(module, self.class_name)
-
-        if not issubclass(check_class, InfrahubCheck):
-            raise ModuleImportError(message=f"The specified class {self.class_name} is not an Infrahub Check")
-
-        return check_class
-
-
-class InfrahubGeneratorDefinitionConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the Generator Definition")
-    file_path: Path = Field(..., description="The file within the repository with the generator code.")
-    query: str = Field(..., description="The GraphQL query to use as input.")
-    parameters: dict[str, Any] = Field(
-        default_factory=dict, description="The input parameters required to run this check"
-    )
-    targets: str = Field(..., description="The group to target when running this generator")
-    class_name: str = Field(default="Generator", description="The name of the generator class to run.")
-    convert_query_response: bool = Field(
-        default=False,
-        description="Decide if the generator should convert the result of the GraphQL query to SDK InfrahubNode objects.",
-    )
-
-    def load_class(
-        self, import_root: Optional[str] = None, relative_path: Optional[str] = None
-    ) -> type[InfrahubGenerator]:
-        module = import_module(module_path=self.file_path, import_root=import_root, relative_path=relative_path)
-
-        if self.class_name not in dir(module):
-            raise ModuleImportError(message=f"The specified class {self.class_name} was not found within the module")
-
-        generator_class = getattr(module, self.class_name)
-
-        if not issubclass(generator_class, InfrahubGenerator):
-            raise ModuleImportError(message=f"The specified class {self.class_name} is not an Infrahub Generator")
-
-        return generator_class
-
-
-class InfrahubPythonTransformConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the Transform")
-    file_path: Path = Field(..., description="The file within the repository with the transform code.")
-    class_name: str = Field(default="Transform", description="The name of the transform class to run.")
-
-    def load_class(
-        self, import_root: Optional[str] = None, relative_path: Optional[str] = None
-    ) -> type[InfrahubTransform]:
-        module = import_module(module_path=self.file_path, import_root=import_root, relative_path=relative_path)
-
-        if self.class_name not in dir(module):
-            raise ModuleImportError(message=f"The specified class {self.class_name} was not found within the module")
-
-        transform_class = getattr(module, self.class_name)
-
-        if not issubclass(transform_class, InfrahubTransform):
-            raise ModuleImportError(message=f"The specified class {self.class_name} is not an Infrahub Transform")
-
-        return transform_class
-
-
-class InfrahubRepositoryGraphQLConfig(InfrahubRepositoryConfigElement):
-    model_config = ConfigDict(extra="forbid")
-    name: str = Field(..., description="The name of the GraphQL Query")
-    file_path: Path = Field(..., description="The file within the repository with the query code.")
-
-    def load_query(self, relative_path: str = ".") -> str:
-        file_name = Path(f"{relative_path}/{self.file_path}")
-        with file_name.open("r", encoding="UTF-8") as file:
-            return file.read()
-
-
-RESOURCE_MAP: dict[Any, str] = {
-    InfrahubJinja2TransformConfig: "jinja2_transforms",
-    InfrahubCheckDefinitionConfig: "check_definitions",
-    InfrahubRepositoryArtifactDefinitionConfig: "artifact_definitions",
-    InfrahubPythonTransformConfig: "python_transforms",
-    InfrahubGeneratorDefinitionConfig: "generator_definitions",
-    InfrahubRepositoryGraphQLConfig: "queries",
-}
-
-
-class InfrahubRepositoryConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-    check_definitions: list[InfrahubCheckDefinitionConfig] = Field(
-        default_factory=list, description="User defined checks"
-    )
-    schemas: list[Path] = Field(default_factory=list, description="Schema files")
-    jinja2_transforms: list[InfrahubJinja2TransformConfig] = Field(
-        default_factory=list, description="Jinja2 data transformations"
-    )
-    artifact_definitions: list[InfrahubRepositoryArtifactDefinitionConfig] = Field(
-        default_factory=list, description="Artifact definitions"
-    )
-    python_transforms: list[InfrahubPythonTransformConfig] = Field(
-        default_factory=list, description="Python data transformations"
-    )
-    generator_definitions: list[InfrahubGeneratorDefinitionConfig] = Field(
-        default_factory=list, description="Generator definitions"
-    )
-    queries: list[InfrahubRepositoryGraphQLConfig] = Field(default_factory=list, description="GraphQL Queries")
-
-    @field_validator(
-        "check_definitions",
-        "jinja2_transforms",
-        "artifact_definitions",
-        "python_transforms",
-        "generator_definitions",
-        "queries",
-    )
-    @classmethod
-    def unique_items(cls, v: list[Any]) -> list[Any]:
-        names = [item.name for item in v]
-        if dups := duplicates(names):
-            raise ValueError(f"Found multiples element with the same names: {dups}")
-        return v
-
-    def _has_resource(self, resource_id: str, resource_type: type[ResourceClass], resource_field: str = "name") -> bool:
-        for item in getattr(self, RESOURCE_MAP[resource_type]):
-            if getattr(item, resource_field) == resource_id:
-                return True
-        return False
-
-    def _get_resource(
-        self, resource_id: str, resource_type: type[ResourceClass], resource_field: str = "name"
-    ) -> ResourceClass:
-        for item in getattr(self, RESOURCE_MAP[resource_type]):
-            if getattr(item, resource_field) == resource_id:
-                return item
-        raise ResourceNotDefinedError(f"Unable to find {resource_id!r} in {RESOURCE_MAP[resource_type]!r}")
-
-    def has_jinja2_transform(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubJinja2TransformConfig)
-
-    def get_jinja2_transform(self, name: str) -> InfrahubJinja2TransformConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubJinja2TransformConfig)
-
-    def has_check_definition(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubCheckDefinitionConfig)
-
-    def get_check_definition(self, name: str) -> InfrahubCheckDefinitionConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubCheckDefinitionConfig)
-
-    def has_artifact_definition(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubRepositoryArtifactDefinitionConfig)
-
-    def get_artifact_definition(self, name: str) -> InfrahubRepositoryArtifactDefinitionConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubRepositoryArtifactDefinitionConfig)
-
-    def has_generator_definition(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubGeneratorDefinitionConfig)
-
-    def get_generator_definition(self, name: str) -> InfrahubGeneratorDefinitionConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubGeneratorDefinitionConfig)
-
-    def has_python_transform(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubPythonTransformConfig)
-
-    def get_python_transform(self, name: str) -> InfrahubPythonTransformConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubPythonTransformConfig)
-
-    def has_query(self, name: str) -> bool:
-        return self._has_resource(resource_id=name, resource_type=InfrahubRepositoryGraphQLConfig)
-
-    def get_query(self, name: str) -> InfrahubRepositoryGraphQLConfig:
-        return self._get_resource(resource_id=name, resource_type=InfrahubRepositoryGraphQLConfig)
-
-
-# ---------------------------------------------------------------------------------
-# Main Infrahub Schema File
-# ---------------------------------------------------------------------------------
-class FilterSchema(BaseModel):
-    name: str
-    kind: str
-    description: Optional[str] = None
-
-
-class RelationshipCardinality(str, Enum):
-    ONE = "one"
-    MANY = "many"
-
-
-class RelationshipDirection(str, Enum):
-    BIDIR = "bidirectional"
-    OUTBOUND = "outbound"
-    INBOUND = "inbound"
-
-
-class BranchSupportType(str, Enum):
-    AWARE = "aware"
-    AGNOSTIC = "agnostic"
-    LOCAL = "local"
-
-
-class RelationshipKind(str, Enum):
-    GENERIC = "Generic"
-    ATTRIBUTE = "Attribute"
-    COMPONENT = "Component"
-    PARENT = "Parent"
-    GROUP = "Group"
-    HIERARCHY = "Hierarchy"
-    PROFILE = "Profile"
 
 
 class DropdownMutation(str, Enum):
@@ -314,205 +80,16 @@ class EnumMutation(str, Enum):
     remove = "SchemaEnumRemove"
 
 
-class SchemaState(str, Enum):
-    PRESENT = "present"
-    ABSENT = "absent"
-
-
-class AttributeSchema(BaseModel):
-    id: Optional[str] = None
-    state: SchemaState = SchemaState.PRESENT
-    name: str
-    kind: str
-    label: Optional[str] = None
-    description: Optional[str] = None
-    default_value: Optional[Any] = None
-    inherited: bool = False
-    unique: bool = False
-    branch: Optional[BranchSupportType] = None
-    optional: bool = False
-    read_only: bool = False
-    choices: Optional[list[dict[str, Any]]] = None
-    enum: Optional[list[Union[str, int]]] = None
-    max_length: Optional[int] = None
-    min_length: Optional[int] = None
-    regex: Optional[str] = None
-    order_weight: Optional[int] = None
-
-
-class RelationshipSchema(BaseModel):
-    id: Optional[str] = None
-    state: SchemaState = SchemaState.PRESENT
-    name: str
-    peer: str
-    direction: RelationshipDirection = RelationshipDirection.BIDIR
-    kind: RelationshipKind = RelationshipKind.GENERIC
-    label: Optional[str] = None
-    description: Optional[str] = None
-    identifier: Optional[str] = None
-    inherited: bool = False
-    cardinality: str = "many"
-    branch: Optional[BranchSupportType] = None
-    optional: bool = True
-    read_only: bool = False
-    filters: list[FilterSchema] = Field(default_factory=list)
-    order_weight: Optional[int] = None
-
-
-class BaseNodeSchema(BaseModel):
-    id: Optional[str] = None
-    state: SchemaState = SchemaState.PRESENT
-    name: str
-    label: Optional[str] = None
-    namespace: str
-    description: Optional[str] = None
-    attributes: list[AttributeSchema] = Field(default_factory=list)
-    relationships: list[RelationshipSchema] = Field(default_factory=list)
-    filters: list[FilterSchema] = Field(default_factory=list)
-
-    @property
-    def kind(self) -> str:
-        return self.namespace + self.name
-
-    def get_field(self, name: str, raise_on_error: bool = True) -> Union[AttributeSchema, RelationshipSchema, None]:
-        if attribute_field := self.get_attribute_or_none(name=name):
-            return attribute_field
-
-        if relationship_field := self.get_relationship_or_none(name=name):
-            return relationship_field
-
-        if not raise_on_error:
-            return None
-
-        raise ValueError(f"Unable to find the field {name}")
-
-    def get_attribute(self, name: str) -> AttributeSchema:
-        for item in self.attributes:
-            if item.name == name:
-                return item
-        raise ValueError(f"Unable to find the attribute {name}")
-
-    def get_attribute_or_none(self, name: str) -> Optional[AttributeSchema]:
-        for item in self.attributes:
-            if item.name == name:
-                return item
-        return None
-
-    def get_relationship(self, name: str) -> RelationshipSchema:
-        for item in self.relationships:
-            if item.name == name:
-                return item
-        raise ValueError(f"Unable to find the relationship {name}")
-
-    def get_relationship_or_none(self, name: str) -> Optional[RelationshipSchema]:
-        for item in self.relationships:
-            if item.name == name:
-                return item
-        return None
-
-    def get_relationship_by_identifier(self, id: str, raise_on_error: bool = True) -> Union[RelationshipSchema, None]:
-        for item in self.relationships:
-            if item.identifier == id:
-                return item
-
-        if not raise_on_error:
-            return None
-
-        raise ValueError(f"Unable to find the relationship {id}")
-
-    def get_matching_relationship(
-        self, id: str, direction: RelationshipDirection = RelationshipDirection.BIDIR
-    ) -> RelationshipSchema:
-        valid_direction = RelationshipDirection.BIDIR
-        if direction == RelationshipDirection.INBOUND:
-            valid_direction = RelationshipDirection.OUTBOUND
-        elif direction == RelationshipDirection.OUTBOUND:
-            valid_direction = RelationshipDirection.INBOUND
-
-        for item in self.relationships:
-            if item.identifier == id and item.direction == valid_direction:
-                return item
-
-        raise ValueError(f"Unable to find the relationship {id} / ({valid_direction.value})")
-
-    @property
-    def attribute_names(self) -> list[str]:
-        return [item.name for item in self.attributes]
-
-    @property
-    def relationship_names(self) -> list[str]:
-        return [item.name for item in self.relationships]
-
-    @property
-    def mandatory_input_names(self) -> list[str]:
-        return self.mandatory_attribute_names + self.mandatory_relationship_names
-
-    @property
-    def mandatory_attribute_names(self) -> list[str]:
-        return [item.name for item in self.attributes if not item.optional and item.default_value is None]
-
-    @property
-    def mandatory_relationship_names(self) -> list[str]:
-        return [item.name for item in self.relationships if not item.optional]
-
-    @property
-    def local_attributes(self) -> list[AttributeSchema]:
-        return [item for item in self.attributes if not item.inherited]
-
-    @property
-    def local_relationships(self) -> list[RelationshipSchema]:
-        return [item for item in self.relationships if not item.inherited]
-
-    @property
-    def unique_attributes(self) -> list[AttributeSchema]:
-        return [item for item in self.attributes if item.unique]
-
-
-class GenericSchema(BaseNodeSchema):
-    """A Generic can be either an Interface or a Union depending if there are some Attributes or Relationships defined."""
-
-    used_by: list[str] = Field(default_factory=list)
-
-
-class NodeSchema(BaseNodeSchema):
-    inherit_from: list[str] = Field(default_factory=list)
-    branch: Optional[BranchSupportType] = None
-    default_filter: Optional[str] = None
-    human_friendly_id: Optional[list[str]] = None
-
-
-class ProfileSchema(BaseNodeSchema):
-    inherit_from: list[str] = Field(default_factory=list)
-
-
-class NodeExtensionSchema(BaseModel):
-    name: Optional[str] = None
-    kind: str
-    description: Optional[str] = None
-    label: Optional[str] = None
-    inherit_from: list[str] = Field(default_factory=list)
-    branch: Optional[BranchSupportType] = None
-    default_filter: Optional[str] = None
-    attributes: list[AttributeSchema] = Field(default_factory=list)
-    relationships: list[RelationshipSchema] = Field(default_factory=list)
-
-
-class SchemaRoot(BaseModel):
-    version: str
-    generics: list[GenericSchema] = Field(default_factory=list)
-    nodes: list[NodeSchema] = Field(default_factory=list)
-    profiles: list[ProfileSchema] = Field(default_factory=list)
-    # node_extensions: list[NodeExtensionSchema] = Field(default_factory=list)
-
-
-MainSchemaTypes: TypeAlias = Union[NodeSchema, GenericSchema, ProfileSchema]
+MainSchemaTypes: TypeAlias = Union[NodeSchema, GenericSchema]
+MainSchemaTypesAPI: TypeAlias = Union[NodeSchemaAPI, GenericSchemaAPI, ProfileSchemaAPI]
+MainSchemaTypesAll: TypeAlias = Union[NodeSchema, GenericSchema, NodeSchemaAPI, GenericSchemaAPI, ProfileSchemaAPI]
 
 
 class InfrahubSchemaBase:
     def validate(self, data: dict[str, Any]) -> None:
         SchemaRoot(**data)
 
-    def validate_data_against_schema(self, schema: MainSchemaTypes, data: dict) -> None:
+    def validate_data_against_schema(self, schema: MainSchemaTypesAPI, data: dict) -> None:
         for key in data.keys():
             if key not in schema.relationship_names + schema.attribute_names:
                 identifier = f"{schema.kind}"
@@ -523,7 +100,7 @@ class InfrahubSchemaBase:
 
     def generate_payload_create(
         self,
-        schema: MainSchemaTypes,
+        schema: MainSchemaTypesAPI,
         data: dict,
         source: Optional[str] = None,
         owner: Optional[str] = None,
@@ -599,7 +176,7 @@ class InfrahubSchema(InfrahubSchemaBase):
         branch: Optional[str] = None,
         refresh: bool = False,
         timeout: Optional[int] = None,
-    ) -> MainSchemaTypes:
+    ) -> MainSchemaTypesAPI:
         branch = branch or self.client.default_branch
 
         kind_str = self._get_schema_name(schema=kind)
@@ -622,7 +199,7 @@ class InfrahubSchema(InfrahubSchemaBase):
 
     async def all(
         self, branch: Optional[str] = None, refresh: bool = False, namespaces: Optional[list[str]] = None
-    ) -> MutableMapping[str, MainSchemaTypes]:
+    ) -> MutableMapping[str, MainSchemaTypesAPI]:
         """Retrieve the entire schema for a given branch.
 
         if present in cache, the schema will be served from the cache, unless refresh is set to True
@@ -809,7 +386,7 @@ class InfrahubSchema(InfrahubSchemaBase):
 
     async def fetch(
         self, branch: str, namespaces: Optional[list[str]] = None, timeout: Optional[int] = None
-    ) -> MutableMapping[str, MainSchemaTypes]:
+    ) -> MutableMapping[str, MainSchemaTypesAPI]:
         """Fetch the schema from the server for a given branch.
 
         Args:
@@ -830,17 +407,17 @@ class InfrahubSchema(InfrahubSchemaBase):
 
         data: MutableMapping[str, Any] = response.json()
 
-        nodes: MutableMapping[str, MainSchemaTypes] = {}
+        nodes: MutableMapping[str, MainSchemaTypesAPI] = {}
         for node_schema in data.get("nodes", []):
-            node = NodeSchema(**node_schema)
+            node = NodeSchemaAPI(**node_schema)
             nodes[node.kind] = node
 
         for generic_schema in data.get("generics", []):
-            generic = GenericSchema(**generic_schema)
+            generic = GenericSchemaAPI(**generic_schema)
             nodes[generic.kind] = generic
 
         for profile_schema in data.get("profiles", []):
-            profile = ProfileSchema(**profile_schema)
+            profile = ProfileSchemaAPI(**profile_schema)
             nodes[profile.kind] = profile
 
         return nodes
@@ -853,7 +430,7 @@ class InfrahubSchemaSync(InfrahubSchemaBase):
 
     def all(
         self, branch: Optional[str] = None, refresh: bool = False, namespaces: Optional[list[str]] = None
-    ) -> MutableMapping[str, MainSchemaTypes]:
+    ) -> MutableMapping[str, MainSchemaTypesAPI]:
         """Retrieve the entire schema for a given branch.
 
         if present in cache, the schema will be served from the cache, unless refresh is set to True
@@ -878,7 +455,7 @@ class InfrahubSchemaSync(InfrahubSchemaBase):
         branch: Optional[str] = None,
         refresh: bool = False,
         timeout: Optional[int] = None,
-    ) -> MainSchemaTypes:
+    ) -> MainSchemaTypesAPI:
         branch = branch or self.client.default_branch
 
         kind_str = self._get_schema_name(schema=kind)
@@ -901,7 +478,7 @@ class InfrahubSchemaSync(InfrahubSchemaBase):
 
     def _get_kind_and_attribute_schema(
         self, kind: Union[str, InfrahubNodeTypes], attribute: str, branch: Optional[str] = None
-    ) -> tuple[str, AttributeSchema]:
+    ) -> tuple[str, AttributeSchemaAPI]:
         node_kind: str = kind._schema.kind if not isinstance(kind, str) else kind
         node_schema = self.client.schema.get(kind=node_kind, branch=branch)
         schema_attr = node_schema.get_attribute(name=attribute)
@@ -1013,7 +590,7 @@ class InfrahubSchemaSync(InfrahubSchemaBase):
 
     def fetch(
         self, branch: str, namespaces: Optional[list[str]] = None, timeout: Optional[int] = None
-    ) -> MutableMapping[str, MainSchemaTypes]:
+    ) -> MutableMapping[str, MainSchemaTypesAPI]:
         """Fetch the schema from the server for a given branch.
 
         Args:
@@ -1034,17 +611,17 @@ class InfrahubSchemaSync(InfrahubSchemaBase):
 
         data: MutableMapping[str, Any] = response.json()
 
-        nodes: MutableMapping[str, MainSchemaTypes] = {}
+        nodes: MutableMapping[str, MainSchemaTypesAPI] = {}
         for node_schema in data.get("nodes", []):
-            node = NodeSchema(**node_schema)
+            node = NodeSchemaAPI(**node_schema)
             nodes[node.kind] = node
 
         for generic_schema in data.get("generics", []):
-            generic = GenericSchema(**generic_schema)
+            generic = GenericSchemaAPI(**generic_schema)
             nodes[generic.kind] = generic
 
         for profile_schema in data.get("profiles", []):
-            profile = ProfileSchema(**profile_schema)
+            profile = ProfileSchemaAPI(**profile_schema)
             nodes[profile.kind] = profile
 
         return nodes
